chore: move useMaxSelection to VaSelect
