+++ packages/ui/src/components/va-select/hooks/useMaxSelections.spec.ts
@@ -1,4 +1,4 @@
-import { useMaxSelections, useMaxSelectionsProps } from '../useMaxSelections'
+import { useMaxSelections, useMaxSelectionsProps } from './useMaxSelections'
 import { mount } from '@vue/test-utils'
 import { ref } from 'vue'
 import { describe, it, expect } from 'vitest'
