+++ packages/ui/src/components/va-select/hooks/useMaxSelections.ts
