+++ packages/ui/src/composables/index.ts
@@ -16,7 +16,6 @@ export * from './useForm'
 export * from './useI18n'
 export * from './std/browser/observer/useIntersectionObserver'
 export * from './std/event/useLongPress'
-export * from './useMaxSelections'
 export * from './useParsableMeasure'
 export * from './std/browser/observer/useResizeObserver'
 export * from './useRouterLink'
